+++ packages/dialect-react-ui/components/common/providers/DialectThemeProvider.tsx
@@ -168,11 +168,11 @@ export const defaultVariables: Record<ThemeType, ThemeValues> &
     // Uses `react-transition-group
     bottomSlide: {
       enter: 'dt-transition-transform dt-duration-300 dt-ease-in-out',
-      enterFrom: '-dt-translate-y-14',
-      enterTo: '-dt-translate-y-full',
+      enterFrom: 'dt-translate-y-[calc(100%-3.5rem)]',
+      enterTo: '!dt-translate-y-0',
       leave: 'dt-transition-transform dt-duration-100 dt-ease-in-out',
-      leaveFrom: '-dt-translate-y-full',
-      leaveTo: '!-dt-translate-y-14',
+      leaveFrom: 'dt-translate-y-0',
+      leaveTo: '!dt-translate-y-[calc(100%-3.5rem)]',
     },
   },
   light: {
@@ -254,7 +254,7 @@ export const defaultVariables: Record<ThemeType, ThemeValues> &
       'dt-fixed dt-z-50 dt-top-0 dt-w-full dt-h-full dt-right-0 sm:dt-absolute sm:dt-top-16 sm:dt-w-[30rem] sm:dt-h-[40rem]',
     modal: 'dt-rounded-none dt-shadow-md sm:dt-rounded-3xl',
     sliderWrapper:
-      'dt-fixed dt-z-50 dt-top-0 dt-w-full dt-h-full sm:dt-w-[30rem] sm:dt-h-[40rem] sm:dt-right-10 sm:dt-top-auto dt-bottom-0',
+      'dt-fixed dt-z-50 dt-top-0 dt-bottom-0 dt-w-full dt-h-full sm:dt-w-[30rem] sm:dt-h-[40rem] sm:dt-right-10 sm:dt-top-auto dt-bottom-0',
     slider: 'dt-rounded-none dt-shadow-md sm:dt-rounded-t-3xl',
     button:
       'dt-bg-black dt-text-white dt-border dt-border-black hover:dt-opacity-60',
@@ -358,7 +358,7 @@ export const defaultVariables: Record<ThemeType, ThemeValues> &
       'dt-fixed dt-z-50 dt-top-0 dt-w-full dt-h-full dt-right-0 sm:dt-absolute sm:dt-top-16 sm:dt-w-[30rem] sm:dt-h-[40rem]',
     modal: 'dt-rounded-none dt-shadow-md sm:dt-rounded-3xl',
     sliderWrapper:
-      'dt-fixed dt-z-50 dt-top-auto dt-w-full dt-h-full sm:dt-w-[30rem] sm:dt-h-[40rem] sm:dt-right-10 sm:dt-top-auto',
+      'dt-fixed dt-z-50 dt-top-auto dt-bottom-0 dt-w-full dt-h-full sm:dt-w-[30rem] sm:dt-h-[40rem] sm:dt-right-10 sm:dt-top-auto',
     slider: 'dt-rounded-none dt-shadow-md sm:dt-rounded-t-3xl',
     button:
       'dt-bg-white dt-text-black dt-border dt-border-white hover:dt-opacity-60',